rename full screen to focus
