+++ tui-module/src/widgets.rs
@@ -1,5 +1,5 @@
 pub mod album_list;
 pub mod artist_list;
-pub mod full_screen;
+pub mod focus;
 pub mod playlist_list;
 pub mod track_list;
